Basic outlining. Tim's code will be merged within the next 12 hours.
I tuned driving mode and also figured out some of the weird bugs I was having with Speed being 0
intermittently. It has to do with printing strings during loop().

Next I'll work on filling out the procedures using the GPS code and the pins for
the metal detector and horn.